fix(init)!: drop `init` command from `@nuxt/cli`
