+++ packages/nuxt-cli/package.json
@@ -68,7 +68,6 @@
     "exsolve": "^1.1.1",
     "fzf": "^0.5.2",
     "get-port-please": "^3.2.0",
-    "giget": "^3.3.1",
     "nypm": "^0.6.8",
     "obug": "^2.1.4",
     "ohash": "^2.0.11",
@@ -92,6 +91,7 @@
     "@nuxt/schema": "^4.5.1",
     "@nuxt/test-utils": "^4.1.0",
     "@types/node": "^24.13.3",
+    "giget": "^3.3.1",
     "h3": "^1.15.11",
     "h3-next": "npm:h3@^2.0.1-rc.26",
     "jiti": "^2.7.0",
